Remove QR decomposition code
We are going to use LAPACKE to do the QR decomposition. So first the
matrix code is rewritten such that it stores the matrices row-major.
Then LAPACK can be used to determine a null space by means of a QR
decomposition. Since we need to determine a QR decomposition of A' for
that, the column-major ordering of LAPACK will help us.