Refactor QR decomposition code
Constructing a QR decomposition is a separate function now. This allows
to solve linear systems using a QR decomposition.

Workflow for constructing a null space will therefore be:
- Cook up a QR decomposition
- Exact rightmost columns from Q

Workflow for solving a linear system will be:
- Cook up a QR decomposition
- Solve linear system using decomposition

Main advantage is that the decomposition can be calculated beforehand,
which is necessary in implremovals.