Replace the memory intensive trie with a tree
The trie used a lot of memory since it used a struct for any integer.
This binary tree implementation only has a small struct for each
removal, which should use significantly less memory.

There are various ways to further improve this implemention. For
example, the binary tree can be balanced to ensure that it has O(N log
N) lookups. A second point can be to further improve the inefficient
comparision done at each tree node.