Provide own implementation of getopt
getopt is not portable (e.g. to Windows) since it is a POSIX symbol.
This is a basic implementation of getopt that fulfills the needs of this
program. It is unlikely that it adheres the POSIX standard, but we also
do not claim that that is the case in the source code.