Remove LAPACK from build options and README
We have our own implementation of the LU factorization now, so LAPACK is
not necessary anymore. This makes compiling the software in operating
systems that do not have built-in support for LAPACK significantly
easier.

Moreover it makes it easier to further extend the linear algebra
routines without having to dig into Fortran.