Do not try to remove nodes with zero vector
To reproduce a removal, we need a vector which has a non-zero entry at
the location of the node that will be removed. Otherwise the null vector
cannot be used to remove the specific node, since it cannot change its
weight.

Fix this by actively searching for the null vector with the absolute
highest value at that position. Such null vector must exist, since the
null space can be used to reproduce the specific removal.