Only support vectors in matrix_lu_solve
The function matrix_lu_solve is a critical function in the code, as it
consumes a large fraction of the running time. A lot of time is consumed
by the most inner loop, that iterates over the number of columns of b.
However, in this code b is *always* a vector, so that inner loop can be
removed. This yields a significant and noticable speed-up.